Force use of nette/security:3.1
The new version (3.2) brings removal of deprecated
IUserStorage and also enforcement of login(Identity)
instead of current array which we use in our
authenticators.

We need to limit the application at the current version,
until we create an identity class allowing us to pass
different types of authentication data to our
authenticators.