Fix tables' self-reference in update clauses
After some improvements in the framework, queries
that used to fetch IDs for subqueries now inject
the subquery itself to the condition.

This caused some issues when doing updates on a table
where we need to execute subquery on the very same
table.

Commit alters the affected places to fetch the IDs
itself so that it mimics the previous implementation.

remp/crm#3407