Fix possibly incomplete user identity after registration
If event handler listening to user registration updated the user,
this change wasn't propagated back to the UserManager which
then returned "original" user to the caller.

This could cause that caller (which is setting up user's
identity) would set it without the attribute changes
made by the listeners. Error should only affect the first
login session and wouldn't be destructive, but stil...

remp/respekt#111