Remove token auth source metadata from UserInfo API handler
The authSource data was an obsolete and problematic
implementation for Nette. If someone needs extra metadata
in the handler, the best approach is to extend/override
it for now.

The IF statement that this commit removes was doing
the problematic part:

  isset($data['token']->authSource)

By doing this, if there was an ActiveRow in the $data['token'],
Nette tried to resolve the foreign key for the "token" column
(which doesn't exist) and effectivelly caused full DB structure
reset, executing queries like this for every table:

  SHOW FULL COLUMNS FROM `access_tokens`;
  SELECT CONSTRAINT_NAME, COLUMN_NAME, REFERENCED_TABLE_NAME, REFERENCED_COLUMN_NAME
  FROM information_schema.KEY_COLUMN_USAGE
  WHERE TABLE_SCHEMA = DATABASE()
    AND REFERENCED_TABLE_NAME IS NOT NULL
    AND TABLE_NAME = 'access_tokens';

During higher loads, these could cause issues.

remp/crm#3663